# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Competitor Monitoring Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Competitor {
  id: number;
  name: string;
  url: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  faviconUrl?: string | null;
  monitorPricing: boolean;
  monitorFeatures: boolean;
  monitorBlog: boolean;
  monitorJobs: boolean;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  lastCheckedAt: string | null;
  /** @nullable */
  checksCount?: number | null;
  /** @nullable */
  changesCount?: number | null;
}

export interface CompetitorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  url: string;
  description?: string;
  monitorPricing?: boolean;
  monitorFeatures?: boolean;
  monitorBlog?: boolean;
  monitorJobs?: boolean;
}

export interface CompetitorUpdate {
  /** @minLength 1 */
  name?: string;
  url?: string;
  description?: string;
  monitorPricing?: boolean;
  monitorFeatures?: boolean;
  monitorBlog?: boolean;
  monitorJobs?: boolean;
  isActive?: boolean;
}

export type CheckStatus = typeof CheckStatus[keyof typeof CheckStatus];


export const CheckStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type CheckChangeType = typeof CheckChangeType[keyof typeof CheckChangeType] | null;


export const CheckChangeType = {
  pricing: 'pricing',
  features: 'features',
  blog: 'blog',
  jobs: 'jobs',
} as const;

export interface Check {
  id: number;
  competitorId: number;
  competitorName: string;
  status: CheckStatus;
  checkedAt: string;
  hasChanges: boolean;
  /** @nullable */
  changeType: CheckChangeType;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export type DigestStatus = typeof DigestStatus[keyof typeof DigestStatus];


export const DigestStatus = {
  sent: 'sent',
  failed: 'failed',
  skipped: 'skipped',
} as const;

export interface Digest {
  id: number;
  sentAt: string;
  status: DigestStatus;
  competitorsChecked: number;
  changesFound: number;
  /** @nullable */
  slackMessageId?: string | null;
  /** @nullable */
  content?: string | null;
}

export type SettingsDigestSchedule = typeof SettingsDigestSchedule[keyof typeof SettingsDigestSchedule];


export const SettingsDigestSchedule = {
  daily: 'daily',
  weekly: 'weekly',
} as const;

export interface Settings {
  id: number;
  slackWebhookConfigured: boolean;
  digestSchedule: SettingsDigestSchedule;
  digestTime: string;
  isDigestEnabled: boolean;
  timezone?: string;
}

export type SettingsUpdateDigestSchedule = typeof SettingsUpdateDigestSchedule[keyof typeof SettingsUpdateDigestSchedule];


export const SettingsUpdateDigestSchedule = {
  daily: 'daily',
  weekly: 'weekly',
} as const;

export interface SettingsUpdate {
  slackWebhookUrl?: string;
  digestSchedule?: SettingsUpdateDigestSchedule;
  digestTime?: string;
  isDigestEnabled?: boolean;
  timezone?: string;
}

export type DashboardSummaryChangesByType = {
  pricing?: number;
  features?: number;
  blog?: number;
  jobs?: number;
};

export interface DashboardSummary {
  totalCompetitors: number;
  activeCompetitors: number;
  totalChecks: number;
  checksToday: number;
  totalChanges: number;
  changesThisWeek: number;
  /** @nullable */
  lastDigestSentAt: string | null;
  digestsThisMonth: number;
  changesByType?: DashboardSummaryChangesByType;
}

export type ActivityItemChangeType = typeof ActivityItemChangeType[keyof typeof ActivityItemChangeType];


export const ActivityItemChangeType = {
  pricing: 'pricing',
  features: 'features',
  blog: 'blog',
  jobs: 'jobs',
} as const;

export interface ActivityItem {
  id: number;
  competitorId: number;
  competitorName: string;
  competitorUrl?: string;
  changeType: ActivityItemChangeType;
  summary: string;
  /** @nullable */
  details?: string | null;
  detectedAt: string;
}

export type ListChecksParams = {
limit?: number;
};

export type ListActivityParams = {
limit?: number;
};