# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Competitor Monitoring Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemChangeType } from './activityItemChangeType';

export interface ActivityItem {
  id: number;
  competitorId: number;
  competitorName: string;
  competitorUrl?: string;
  changeType: ActivityItemChangeType;
  summary: string;
  /** @nullable */
  details?: string | null;
  detectedAt: string;
}