# lib/api-zod/src/generated/types/competitor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Competitor Monitoring Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Competitor {
  id: number;
  name: string;
  url: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  faviconUrl?: string | null;
  monitorPricing: boolean;
  monitorFeatures: boolean;
  monitorBlog: boolean;
  monitorJobs: boolean;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  lastCheckedAt: string | null;
  /** @nullable */
  checksCount?: number | null;
  /** @nullable */
  changesCount?: number | null;
}