# lib/api-zod/src/generated/types/competitorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Competitor Monitoring Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CompetitorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  url: string;
  description?: string;
  monitorPricing?: boolean;
  monitorFeatures?: boolean;
  monitorBlog?: boolean;
  monitorJobs?: boolean;
}