# lib/api-zod/src/generated/types/competitorUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Competitor Monitoring Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CompetitorUpdate {
  /** @minLength 1 */
  name?: string;
  url?: string;
  description?: string;
  monitorPricing?: boolean;
  monitorFeatures?: boolean;
  monitorBlog?: boolean;
  monitorJobs?: boolean;
  isActive?: boolean;
}