# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Competitor Monitoring Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryChangesByType } from './dashboardSummaryChangesByType';

export interface DashboardSummary {
  totalCompetitors: number;
  activeCompetitors: number;
  totalChecks: number;
  checksToday: number;
  totalChanges: number;
  changesThisWeek: number;
  /** @nullable */
  lastDigestSentAt: string | null;
  digestsThisMonth: number;
  changesByType?: DashboardSummaryChangesByType;
}