# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Competitor Monitoring Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsDigestSchedule } from './settingsDigestSchedule';

export interface Settings {
  id: number;
  slackWebhookConfigured: boolean;
  digestSchedule: SettingsDigestSchedule;
  digestTime: string;
  isDigestEnabled: boolean;
  timezone?: string;
}