# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Competitor Monitoring Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsUpdateDigestSchedule } from './settingsUpdateDigestSchedule';

export interface SettingsUpdate {
  slackWebhookUrl?: string;
  digestSchedule?: SettingsUpdateDigestSchedule;
  digestTime?: string;
  isDigestEnabled?: boolean;
  timezone?: string;
}